import time fixes
