+++ weave/type_handlers/File/file.py
@@ -15,8 +15,6 @@ from typing_extensions import deprecated
 
 from weave.trace.serialization import serializer
 from weave.trace.serialization.custom_objs import MemTraceFilesArtifact
-from weave.type_handlers.Content.content import save as save_content
-from weave.type_wrappers.Content.content import Content
 
 logger = logging.getLogger(__name__)
 
@@ -93,6 +91,7 @@ class File:
 
 @deprecated(DEPRECATION_MESSAGE)
 def save(obj: File, artifact: MemTraceFilesArtifact, name: str) -> None:
+    from weave.type_handlers.Content.content import Content, save as save_content
     logger.warning("Saving File as Content object")
     content: Content = Content.from_path(obj.path, mimetype=obj.mimetype)
     save_content(content, artifact, name)
